Seperate things into their own files, move everything around
